Publish media codec profiling measurements
This exposes the VideoCapabilities.getAchievableFrameRatesFor
interface for app developers to utilise to get the range
of achievable frame rates for a video size.

This also fixes the CTS tests for:
com.android.cts.videoperf.VideoEncoderDecoderTest
android.media.cts.VideoDecoderPerfTest

Note that these are pretty rough and preliminary measurements,
and they could be further improved in the future.

Change-Id: Ic392c20d9f521ad5e98104ad0cecec89fe2d7006