init: Move perfd directory creation to post-fs-data
When a device is encrypted, its /data partition is remounted after
decryption. As the perfd directories are created only on boot, they
will no longer be present when the /data partition is remounted.

To fix this, create the perfd directories on post-fs-data. This ensures
that the directories are recreated after decryption.

Change-Id: I604921cc09384109f6190ef0d34406b3a145ea8a