Set the actual minimum brightness allowed by panel
The kernel driver enforces a minimum display brightness value of 10.
Advertising support for a brightness < 10 gives the user a false
idea of the range of brightness values.

Hence, set all minimum brightness values to 10.
config_screenBrightnessSettingMinimum and config_screenBrightnessDim
need not be set as the default values for these configs are already 10.

Change-Id: I6f0a5a01cd4e88ec9c8e688091d68e2ae70e2647